requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.16
django-mongoengine==0.5.6
djangorestframework==3.15.2
dnspython==2.6.1
mongoengine==0.29.0
pymongo==4.9.1
sqlparse==0.5.1
typing_extensions==4.12.2
